Add protobuf_path and dataset_type arguments
To avoid having to modify protobuf_path and data_root_path in the code.
This is useful for instance when generating bags with a script.
Meaning of the arguments added:
- protobuf_path: path to the protobuf file. Simply made it an argument
rather than an hard-coded variable.
- dataset_type: either 'val' or 'train', indicates the type of the
pySceneNetRGBD from which to generate the bag. It is used to construct
data_root_path.